BUG/MINOR: checks: don't send proxy protocol with agent checks
James Brown reported that agent-check mistakenly sends the proxy
protocol header when it's configured. This is obviously wrong as
the agent is an independant servie and not a traffic port, let's
disable this.

This fix must be backported to 1.7 and possibly 1.6.

(cherry picked from commit f494977bc1a361c26f8cc0516366ef2662ac9502)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>