BUG/MEDIUM: cfgparse: Check if tune.http.maxhdr is in the range 1..32767
We cannot store more than 32K headers in the structure hdr_idx, because
internaly we use signed short integers. To avoid any bugs (due to an integers
overflow), a check has been added on tune.http.maxhdr to be sure to not set a
value greater than 32767 and lower than 1 (because this is a nonsense to set
this parameter to a value <= 0).

The documentation has been updated accordingly.

This patch can be backported in 1.7, 1.6 and 1.5.

[wla: bug raised to MEDIUM since it fail with a new error at startup]
(cherry picked from commit 50174f36003141297a4824dd0b98a2f0e6d4997d)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>