BUG/MINOR: http: Don't reset the transaction if there are still data to send
To reset an HTTP transaction, we need to be sure all data were sent, for the
request and the response. There are tests on request and response buffers for
that in http_resync_states function. But the return code was wrong. We must
return 0 to wait.

This patch must be backported in 1.7

(cherry picked from commit a94fda30bd5638a77ea96a1c5fca7945fe736714)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>