BUG/MEDIUM: connection: remove useless flag CO_FL_DATA_RD_SH
This flag is both confusing and wrong. It is supposed to report the
fact that the data layer has received a shutdown, but in fact this is
reported by CO_FL_SOCK_RD_SH which is set by the transport layer after
this condition is detected. The only case where the flag above is set
is in the stream interface where CF_SHUTR is also set on the receiving
channel.

In addition, it was checked in the health checks code (while never set)
and was always test jointly with CO_FL_SOCK_RD_SH everywhere, except in
conn_data_read0_pending() which incorrectly doesn't match the second
time it's called and is fortunately protected by an extra check on
(ic->flags & CF_SHUTR).

This patch gets rid of the flag completely. Now conn_data_read0_pending()
accurately reports the fact that the transport layer has detected the end
of the stream, regardless of the fact that this state was already consumed,
and the stream interface watches ic->flags&CF_SHUTR to know if the channel
was already closed by the upper layer (which it already used to do).

The now unused conn_data_read0() function was removed.
(cherry picked from commit 54e917cfa1e7b0539550ae32c48c76da2f169041)

[wt: this happens to fix a real bug which occasionally strikes when
     using http-reuse in the rare case where a server shuts down after
     providing its response but before the connection is put back into
     the idle pool, and it gets immediately recycled for another request,
     without first passing through the idle handler, and the already
     reported shutdown is never reported to the second transaction,
     causing a loop to last for as long as the server timeout]