BUG/MINOR: spoe: Don't compare engine name and SPOE scope when both are NULL
SPOE filter can be declared without engine name. This is an optional
parameter. But in this case, no scope must be used in the SPOE configuration
file. So engine name and scope are both undefined, and, obviously, we must not
try to compare them.

This patch must be backported in 1.7.

(cherry picked from commit e1405e5f802123eec8270ab87d6e9637c40dfd6a)
Signed-off-by: Willy Tarreau <w@1wt.eu>