BUG/MINOR: tcp-check: don't quit with pending data in the send buffer
In the rare case where the "tcp-check send" directive is the last one in
the list, it leaves the loop without sending the data. Fortunately, the
polling is still enabled on output, resulting in the connection handler
calling back to send what remains, but this is ugly and not very reliable.

This may be backported to 1.7 and 1.6.

(cherry picked from commit 82feaaf0423919c5af61b59f6f3234bc0423a791)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>