BUG/MINOR: ssl: ocsp response with 'revoked' status is correct
ocsp_status can be 'good', 'revoked', or 'unknown'. 'revoked' status
is a correct status and should not be dropped.
In case of certificate with OCSP must-stapling extension, response with
'revoked' status must be provided as well as 'good' status.

This patch can be backported in 1.7, 1.6 and 1.5.

(cherry picked from commit 872085ce375c76723ff023d2a79e9d9a0852aa98)
Signed-off-by: Willy Tarreau <w@1wt.eu>