BUG/MEDIUM: cache: bad computation of the remaining size
The cache was not setting the hdrs_len to zero when we are called
in the http_forward_data with headers + body.

The consequence is to always try to store a size - the size of headers,
during the calls to http_forward_data even when it has already forwarded
the headers.

Thanks to Cyril Bonté for reporting this bug.

Must be backported to 1.8.
(cherry picked from commit bcd9101a6652856d2db30b254906a2c5eb1bb20a)

Signed-off-by: Willy Tarreau <w@1wt.eu>