BUG/MINOR: h2: immediately close if receiving GOAWAY after the last stream
The h2spec test suite reveals that a GOAWAY frame received after the
last stream doesn't cause an immediate close, because we count on the
last stream to quit to do so. By simply setting the last_sid to the
received value in case it was not set, we can ensure to properly close
an idle connection during h2_wake().

To be backported to 1.8.

(cherry picked from commit 11cc2d6031a8fefe78f083c4d364c8489b7dbbaf)
Signed-off-by: Willy Tarreau <w@1wt.eu>