BUG/MINOR: h2: try to abort closed streams as soon as possible
The purpose here is to be able to signal receipt of RST_STREAM to
streams when they start to provide a response so that the response
can be aborted ASAP. Given that RST_STREAM immediately switches the
stream to the CLOSED state, we must check for CLOSED in addition to
the existing ERROR check.

To be backported to 1.8.

(cherry picked from commit 9470d2cd35938db8e053f390c5de0e94eefe254b)
Signed-off-by: Willy Tarreau <w@1wt.eu>