BUG/MINOR: action: Don't check http capture rules when no id is defined
This is a regression in the commit 29730ba5 ("MINOR: action: Add a functions to
check http capture rules"). We must check the capture id only when an id is
defined.

This patch must be backported in 1.8.
(cherry picked from commit fd608dd2d212928e585da645d7d0c7089424a0de)

Signed-off-by: Willy Tarreau <w@1wt.eu>